+++ include/utils/vector.h
@@ -12,13 +12,13 @@ enum Type_Of_Items {
 
 struct Vector {
     void** data;
-    unsigned long long size;
-    unsigned long long capacity;
+    int size;
+    int capacity;
     enum Type_Of_Items type;
 };
 
 
-struct Vector* init_vector(unsigned long long capacity, enum Type_Of_Items);
+struct Vector* init_vector(int capacity, enum Type_Of_Items);
 
 
 void push_back_token(struct Vector* vector, struct Token* token);
